requirements.txt
Flask==1.1.1
google-cloud-pubsub==1.3.1
requests==2.23.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
pprint==0.1
python-firebase==1.2
urllib3==1.25.8
Werkzeug==1.0.0